backend/requirements.txt
fastapi
uvicorn
pydantic
requests
spacy
beautifulsoup4
requests
lxml
python-dotenv
